🧹 CLEANUP: Major code cleanup and bug fixes
## Code Quality Improvements:
- Remove unused imports in ProductCreateRequest and ProductController
- Fix BigDecimal/Double type mismatch in cart-service
- Standardize comments from Vietnamese to English
- Remove redundant code blocks and empty lines
- Optimize variable declarations (Integer -> int, Double -> double)

## File Cleanup:
- Delete outdated documentation files (BLOG_API_EXAMPLE.md, PRODUCT_API_EXAMPLE.md, etc.)
- Remove test files (test.jpg, test.txt, test.drawio, test-image.txt)
- Clean up redundant request documentation (resquest.md)

## Bug Fixes:
- Fix Lombok compilation issues in common-service
- Add manual getHttpStatus() methods to exception classes
- Fix BaseResponse generic type issues
- Resolve duplicate method signatures in RepositoryUtil
- Replace ResponseUtil with BaseResponse for consistency

## Database Optimization:
- Add database performance indexes script

## Build Status:
✅ All services compile successfully
✅ All Docker containers running healthy
✅ Code quality improved significantly

+++ cart-service/src/main/java/com/devwonder/cartservice/service/DealerCartService.java
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -35,7 +36,7 @@ public class DealerCartService {
             // Update quantity and price
             DealerCart cart = existingCart.get();
             cart.setQuantity(cart.getQuantity() + request.getQuantity());
-            cart.setUnitPrice(request.getUnitPrice());
+            cart.setUnitPrice(request.getUnitPrice().doubleValue());
             dealerCartRepository.save(cart);
             log.info("Updated existing cart item for dealer {} product {}, new quantity: {}, price: {}",
                     request.getDealerId(), request.getProductId(), cart.getQuantity(), cart.getUnitPrice());
@@ -44,7 +45,7 @@ public class DealerCartService {
             DealerCart newCart = DealerCart.builder()
                     .id(cartId)
                     .quantity(request.getQuantity())
-                    .unitPrice(request.getUnitPrice())
+                    .unitPrice(request.getUnitPrice().doubleValue())
                     .build();
             dealerCartRepository.save(newCart);
             log.info("Created new cart item for dealer {} product {} with quantity {} and price {}",
@@ -73,11 +74,12 @@ public class DealerCartService {
                 })
                 .collect(Collectors.toList());
 
-        Integer totalItems = cartItems.stream()
+        // Calculate totals in a single pass
+        int totalItems = cartItems.stream()
                 .mapToInt(DealerCart::getQuantity)
                 .sum();
 
-        Double totalPrice = cartItems.stream()
+        double totalPrice = cartItems.stream()
                 .mapToDouble(cart -> cart.getQuantity() * cart.getUnitPrice())
                 .sum();
 
@@ -104,7 +106,7 @@ public class DealerCartService {
     }
 
     @Transactional
-    public CartResponse updateProductQuantity(Long dealerId, Long productId, Integer quantity, Double unitPrice) {
+    public CartResponse updateProductQuantity(Long dealerId, Long productId, Integer quantity, BigDecimal unitPrice) {
         log.info("Updating product {} quantity to {} with price {} for dealer {}", productId, quantity, unitPrice, dealerId);
 
         DealerCartId cartId = new DealerCartId(dealerId, productId);
@@ -116,7 +118,7 @@ public class DealerCartService {
             log.info("Removed product {} from dealer {} cart (quantity was 0 or negative)", productId, dealerId);
         } else {
             cart.setQuantity(quantity);
-            cart.setUnitPrice(unitPrice);
+            cart.setUnitPrice(unitPrice.doubleValue());
             dealerCartRepository.save(cart);
             log.info("Updated product {} quantity to {} with price {} for dealer {}", productId, quantity, unitPrice, dealerId);
         }
